backend/requirements.txt
akshare==1.16.43
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bcrypt==4.0.1
beautifulsoup4==4.13.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click>=8.0.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crypto==1.4.1
cryptography==44.0.2
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator>=1.1.3
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi>=0.68.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.0.0
jiter==0.9.0
joblib==1.4.2
jsonpath==0.82.2
kombu==5.5.0
lightgbm==4.1.0
lxml==5.3.1
mini-racer==0.12.4
Naked==0.1.32
numpy>=1.21.2
openai==1.12.0
openpyxl==3.1.5
packaging==24.2
pandas>=1.3.3
passlib[bcrypt]>=1.7.4
pluggy==1.5.0
prompt_toolkit==3.0.50
py-mini-racer==0.6.0
pyasn1==0.4.8
pycparser==2.22
pydantic>=1.8.2
pydantic-settings==2.1.0
pydantic_core==2.16.2
pypinyin==0.53.0
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
scikit-learn>=0.24.2
scipy==1.13.1
shellescape==3.8.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy>=1.4.23
starlette==0.38.6
tabulate==0.9.0
threadpoolctl==3.5.0
tomli==2.2.1
tqdm==4.67.1
tushare==1.2.89
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn>=0.15.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==0.57.0
xgboost==2.0.3
xlrd==2.0.1
yfinance>=0.1.63
